mapstruct implementations with change in structure of dtos.

+++ src/main/java/OrderManager/Domain/Model/Order.java
@@ -44,7 +44,7 @@ public class Order {
     private List<OrderItem> items;
 
     private BigDecimal subTotal;
-    private double deliveryFee;
+    private BigDecimal deliveryFee;
     private BigDecimal tax;
     private BigDecimal totalPrice;
     private String notes;
